fix: apply code review suggestions cy-89
