+++ apps/frontend/src/assets/css/variables/layout.css
@@ -3,5 +3,4 @@
 	--size-text-max-width: 900px;
 	--app-header-height: 64px;
 	--sidebar-width: 250px;
-	--sidebar-border-width: 3px;
 }
