+++ apps/frontend/src/libs/components/dashboard-wrapper/styles.module.css
@@ -1,11 +1,18 @@
-.appContainer {
+.app-container {
 	min-height: 100vh;
 }
 
-.mainLayout {
+.main-layout {
 	display: flex;
 	height: 100vh;
 	padding-top: var(--app-header-height);
 	overflow: hidden;
 	transition: all 0.3s ease-in-out;
+}
+
+.main-content {
+	flex: 1;
+	min-width: 0;
+	overflow-x: auto;
+	transition: padding 0.3s ease-in-out;
 }
