--- apps/frontend/src/libs/components/dashboard/components/button/styles.module.css
@@ -1,163 +1,0 @@
-.button {
-	--button-padding-block: var(--space-xs);
-	--button-padding-inline: var(--space-2xl);
-	--button-bg-color: var(--color-brand-primary);
-	--button-text-color: var(--color-bg-dark);
-	--button-border-color: var(--color-bg-dark);
-	--button-font-size: var(--text-button);
-
-	position: relative;
-	padding: var(--button-padding-block) var(--button-padding-inline);
-	font-size: var(--button-font-size);
-	font-weight: var(--font-weight-button);
-	color: var(--button-text-color);
-	text-transform: uppercase;
-	background-color: var(--button-bg-color);
-	border: 2px solid var(--button-border-color);
-	border-radius: var(--radius-full);
-	box-shadow: 0 4px 8px #0000001a;
-	transform: translateY(0);
-	transition: all 0.1s cubic-bezier(0.4, 0, 0.2, 1);
-}
-
-.button-small {
-	--button-padding-block: var(--space-2xs);
-	--button-padding-inline: var(--space-xl);
-	--button-font-size: var(--text-button-small);
-}
-
-.button-cluster {
-	--cluster-direction: row;
-	--gutter: var(--space-xs);
-	--cluster-vertical-alignment: center;
-	--cluster-horizontal-alignment: center;
-}
-
-.button:disabled {
-	cursor: not-allowed;
-	opacity: 0.5;
-}
-
-.button:not(:disabled):hover {
-	--button-bg-color: var(--color-bg-dark);
-	--button-text-color: var(--color-bg);
-
-	box-shadow: 0 6px 16px #00000033;
-	transform: translateY(-2px);
-}
-
-.button:not(:disabled):active {
-	box-shadow: 0 2px 8px #00000026;
-	transform: translateY(0);
-	transition: all 0.15s ease;
-}
-
-.button-primary {
-	--button-bg-color: var(--color-brand-primary);
-	--button-border-color: var(--color-bg-dark);
-}
-
-.button-secondary {
-	--button-bg-color: transparent;
-	--button-border-color: var(--color-bg-dark);
-}
-
-.button-secondary:not(:disabled):hover {
-	--button-bg-color: var(--color-brand-primary);
-	--button-text-color: var(--color-bg-dark);
-}
-
-.button-transparent {
-	--button-bg-color: transparent;
-	--button-border-color: transparent;
-}
-
-.button-transparent:not(:disabled):hover {
-	--button-bg-color: var(--color-brand-primary);
-	--button-text-color: var(--color-bg-dark);
-}
-
-.button-icon {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	transition: transform 0.3s ease;
-}
-
-.button:not(:disabled):hover .button-icon {
-	transform: scale(1.2);
-}
-
-.button-icon svg {
-	width: 1.25lh;
-	height: 1.25lh;
-	transition: transform 0.3s ease;
-}
-
-.button-icon-only {
-	--button-icon-only-padding: var(--space-xs);
-
-	aspect-ratio: 1;
-	padding: var(--button-icon-only-padding);
-	background-color: var(--color-bg-light);
-	border: 0;
-	border-radius: var(--radius-full);
-	transform: scale(1);
-	transition: all 0.1s cubic-bezier(0.4, 0, 0.2, 1);
-}
-
-.button-icon-only svg {
-	--icon-size: 1.5lh;
-
-	width: var(--icon-size);
-	height: var(--icon-size);
-}
-
-.button-icon-only:not(:disabled):hover {
-	background-color: var(--color-bg-dark);
-	box-shadow: 0 4px 12px #00000026;
-}
-
-.button-icon-only-small {
-	--button-icon-only-padding: var(--space-2xs);
-}
-
-.button-icon-only-small svg {
-	--icon-size: 1.2lh;
-}
-
-.button-icon-only-medium {
-	--button-icon-only-padding: var(--space-xs);
-}
-
-.button-icon-only-large {
-	--button-icon-only-padding: var(--space-s);
-}
-
-.button-icon-only-large svg {
-	--icon-size: 2lh;
-}
-
-.button:not(.button-small, .button-transparent, .button-icon-only)::after {
-	position: absolute;
-	top: -11px;
-	right: -11px;
-	width: var(--space-l);
-	height: var(--space-l);
-	content: "";
-	border: 2px solid var(--color-bg-dark);
-	border-bottom: none;
-	border-left: none;
-	border-radius: 0 var(--radius-full) 0 0;
-	transform: scale(1);
-	transition: all 0.3s ease;
-}
-
-.button:not(
-		:disabled,
-		.button-small,
-		.button-transparent,
-		.button-icon-only
-	):hover::after {
-	transform: scale(1.1) rotate(8deg);
-}
