+++ apps/frontend/src/libs/components/dashboard/components/greeting/greeting.tsx
@@ -3,7 +3,7 @@ import { type FC } from "react";
 import { CreateDoc } from "~/assets/img/icons/icons.js";
 import { useAppSelector } from "~/libs/hooks/hooks.js";
 
-import { Button } from "../button/button.js";
+import { CreateButton } from "../create-button/create-button.js";
 import styles from "./styles.module.css";
 
 const Greeting: FC = () => {
@@ -12,13 +12,11 @@ const Greeting: FC = () => {
 	return (
 		<div className={styles["container"]}>
 			<h2 className={styles["greeting"]}>Hey {user?.name} 👋</h2>
-			<Button
-				className={styles["button"]}
+			<CreateButton
+				className={styles["button"] ?? ""}
 				disabled
-				icon={<CreateDoc className={styles["largeIcon"]} />}
+				icon={<CreateDoc className={styles["large-icon"]} />}
 				label="Create New Plan"
-				size="large"
-				variant="primary"
 			/>
 		</div>
 	);
