+++ apps/frontend/src/libs/components/dashboard/styles.module.css
@@ -1,7 +1,6 @@
-.mainContent {
-	flex: 1;
-	min-width: 0;
-	padding: var(--space-l);
+.content {
+	height: 100%;
+	padding: var(--space-m);
 	overflow-x: auto;
 	transition: padding 0.3s ease-in-out;
 }
@@ -11,21 +10,21 @@
 	--grid-line: 1px;
 }
 
-.contentGrid {
+.content-grid {
 	display: grid;
-	grid-template-columns: 1fr 1fr;
+	grid-template-columns: 1fr;
 	gap: var(--space-l);
 	transition: gap 0.3s ease-in-out;
 }
 
-@media (width <= 1024px) {
-	.contentGrid {
-		grid-template-columns: 1fr;
+@media (width >= 768px) {
+	.content {
+		padding: var(--space-l);
 	}
 }
 
-@media (width <= 768px) {
-	.mainContent {
-		padding: var(--space-m);
+@media (width >= 1024px) {
+	.content-grid {
+		grid-template-columns: 1fr 1fr;
 	}
 }
