+++ apps/frontend/src/pages/dashboard-wrapper-mock/components/plan/styles.module.css
@@ -2,6 +2,7 @@
 	display: flex;
 	flex: 1;
 	flex-direction: column;
+	height: 100%;
 }
 
 .nav {
